pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archon-trading-ecosystem"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  # Core packages
  "numpy>=1.26",
  "pandas>=2.2",
  "polars>=1.4",
  "pyarrow>=15",
  "pyyaml>=6.0",

  # Visualization
  "matplotlib>=3.7",
  "plotly>=5.15",
  "seaborn>=0.12",

  # Trading & Analysis
  "vectorbt>=0.26",
  "polygon-api-client>=1.12",
  "requests>=2.31",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "scikit-optimize>=0.9",

  # Web & Interface
  "streamlit>=1.36",
  "fastapi>=0.104",
  "websockets>=12.0",
  "uvicorn>=0.24",

  # AI & Knowledge
  "openai>=1.0",
  "chromadb>=0.4",  # Vector database
  "sentence-transformers>=2.2",  # Embeddings
  "pypdf>=3.17",  # PDF processing

  # Utilities
  "rich>=13",
  "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "knowledge*", "trading_engine*", "interfaces*", "core*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"